Rename: ApiResponseLong 위치 변경 (remote.model.response.community -> remote.model.response)
