+++ core/remote/src/main/java/kr/co/remote/CommunityRemoteDataSourceImpl.kt
@@ -9,12 +9,13 @@ import io.ktor.client.request.put
 import io.ktor.client.request.setBody
 import kr.co.data.source.remote.CommunityRemoteDataSource
 import kr.co.remote.model.request.community.BulletinReqDto
-import kr.co.remote.model.response.community.ApiResponseLong
+import kr.co.remote.model.response.ApiResponseLong
 import kr.co.remote.model.response.community.DeleteBulletinResponse
 import kr.co.remote.model.response.community.GetBulletinDetailResponse
 import kr.co.remote.model.response.community.GetBulletinsResponse
 import kr.co.remote.model.response.community.convertToData
 import kr.co.remote.model.response.community.convertToResult
+import kr.co.remote.model.response.convertToData
 import javax.inject.Inject
 
 internal class CommunityRemoteDataSourceImpl @Inject constructor(
